Fixed linking errors.
