+++ SPID/cmake/sourcelist.cmake
@@ -3,7 +3,6 @@ set(sources ${sources}
 	src/Distribute.cpp
 	src/DistributePCLevelMult.cpp
 	src/KeywordDependencies.cpp
-	src/LogBuffer.cpp
 	src/LookupConfigs.cpp
 	src/LookupFilters.cpp
 	src/LookupForms.cpp
